Move function to more appropriate file
